fixed add post

+++ controllers/auth.controller.js
@@ -68,15 +68,20 @@ authController.login = (req, res) => {
           { expiresIn: "1h" }
         );
 
+        // Save auth model in the database
         const auth = new AuthModel({
           user: user._id,
-          token,
-          expiresIn: 3600,
+          token: token,
+          expiresIn: 3600000, // 1 hour in milliseconds
         });
 
         auth
           .save()
           .then(() => {
+            res.cookie("auth-token", token, {
+              maxAge: 3600000,
+              httpOnly: true,
+            });
             res.status(200).json({ token });
           })
           .catch((error) => {
@@ -91,7 +96,11 @@ authController.login = (req, res) => {
 
 // Check if user is authenticated
 authController.isAuthenticated = (req, res, next) => {
-  const token = req.headers.authorization;
+  const token = req.cookies["auth-token"]; // Access 'auth-token' cookie
+
+  if (!token) {
+    return res.status(401).json({ error: "Invalid token" });
+  }
 
   jwt.verify(token, config.secret, (err, decoded) => {
     if (err) {
@@ -103,13 +112,16 @@ authController.isAuthenticated = (req, res, next) => {
     AuthModel.findOne({ token })
       .then((auth) => {
         if (!auth) {
-          return res.status(401).json({ error: "Invalid token" });
+          return res
+            .status(401)
+            .json({ error: "Token doesnt exist in database" });
         }
 
         req.userId = userId;
         next();
       })
       .catch((error) => {
+        console.error("Error saving post:", error);
         res.status(500).json({ error: "Internal server error" });
       });
   });
@@ -117,13 +129,18 @@ authController.isAuthenticated = (req, res, next) => {
 
 // User logout and revoke JWT token
 authController.logout = (req, res) => {
-  const token = req.headers.authorization;
+  const token = req.cookies["auth-token"]; // Access 'auth-token' cookie
+
+  if (!token) {
+    return res.status(404).json({ error: "Token not found" });
+  }
 
   AuthModel.findOneAndDelete({ token })
     .then((deletedToken) => {
       if (!deletedToken) {
         return res.status(404).json({ error: "Token not found" });
       }
+      res.clearCookie("auth-token"); // Clear the 'auth-token' cookie
       res.status(200).json({ message: "Logout successful" });
     })
     .catch((error) => {
